Corrections in Gitlab CI: trying to make all0in-one single executable (#111)
* Corrected compilation warnings for windows

* trying to fix all in CLI

* trying to use compiler provided from environment variable (changes for AppImage)

* fixing compilation (hopefully)

* fixing build (hopefully)

* added full path to compiler embedded in appimage

* trying to use full path to embedded compiler

* adding clang .so to appimage: libclang-cpp.so*

* adding real libraries not symlincs to clang.so

* added libLLVM.so to embedded clang

* fix yaml syntax

* corrected path to headers in AppImage

* include directories are not flat in apimage

* Appimage - fighting with c++config.h

* corrected order of includes

* vindows build - trying to add dependencies

* adding glib runtime to AppImage

* added one missing file

* removed Debug build for Windows

* added missing files to linker

* add other missing files for linker in AppImage

* added gitlab flow to cmake

+++ widgets/SceneWidget.cpp
@@ -1031,8 +1031,8 @@ QString SceneWidget::getNodeAtWorldPosition(const std::array<double, 3>& worldPo
     const int nodeY = static_cast<int>((worldPos[1] - bounds[2]) / nodeHeight);
 
     // Check if the calculated node is within bounds
-    if (nodeX >= 0 && nodeX < settingParameter->nNodeX &&
-        nodeY >= 0 && nodeY < settingParameter->nNodeY)
+    if (nodeX >= 0 && nodeX < static_cast<int>(settingParameter->nNodeX) &&
+        nodeY >= 0 && nodeY < static_cast<int>(settingParameter->nNodeY))
     {
         return QString("Node [%1, %2]").arg(nodeX).arg(nodeY);
     }
